moved notifs to its own component
